Refactor xtask into focused modules
